requirements.txt
pydantic>=2.0
openai>=1.0
python-dotenv>=1.0
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
pytest>=7.0
httpx>=0.27
